pipeline/pyproject.toml
[project]
name = "product-graph-pipeline"
version = "0.1.0"
description = "Offline pipeline: ingest, normalize, resolve entities, embed, and seed Supabase"
requires-python = ">=3.11"
dependencies = [
    "supabase>=2.0",
    "duckdb>=1.0",
    "scikit-learn>=1.5",
    "sentence-transformers>=3.0",
    "pandas>=2.2",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
